User and product test

+++ migrations/20200130142317-create-product.js
@@ -16,7 +16,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       price: {
-        type: Sequelize.DECIMAL(10, 2)
+        type: Sequelize.INTEGER
       },
       stock: {
         type: Sequelize.INTEGER
